# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nxs-anomaly native API",
    "version": "0.1.74",
    "description": "The native /api/v1 contract for nxs-anomaly (API, worker, frontend). This is the authoritative contract; the Grafana-compatibility surface under /api/internal/v1 is a compatibility shim and is NOT part of this contract. The core payloads — user, team, schedule, escalation chain, integration, alert group, alert, notification, delivery attempt — are modelled field by field, with `required` reflecting exactly the keys the engine writes unconditionally; every one still allows additional properties because the underlying storage is JSONB. Remaining resources use the open `Entity` schema. Drift between this document and the router is gated by internal/server/openapi_contract_test.go, and the frontend compiles against types generated from this file (frontend/src/api/schema.d.ts).",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (frontend nginx proxies /api to the API service)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyHeader": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The current principal",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Identity"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/preferences": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Save your own UI language and timezone",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserPreferences"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change your own password",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List your active sessions",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ObjectList"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List users",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserList"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/User"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/migrate-notification-policies": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Backfill default notification policies from legacy targets (idempotent)",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List teams",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TeamList"
          }
        }
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create a team",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Team"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/schedules": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "List schedules",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ScheduleList"
          }
        }
      },
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Create a schedule",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Schedule"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/escalation-chains": {
      "get": {
        "tags": [
          "escalation-chains"
        ],
        "summary": "List escalation chains",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/EscalationChainList"
          }
        }
      },
      "post": {
        "tags": [
          "escalation-chains"
        ],
        "summary": "Create an escalation chain",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/EscalationChain"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/maintenance-windows": {
      "get": {
        "tags": [
          "maintenance-windows"
        ],
        "summary": "List maintenance windows",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MaintenanceWindowList"
          }
        }
      },
      "post": {
        "tags": [
          "maintenance-windows"
        ],
        "summary": "Plan a maintenance window",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/MaintenanceWindow"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/integrations": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List integrations",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/IntegrationList"
          }
        }
      },
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Create an integration",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Integration"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/chatops/channels": {
      "get": {
        "tags": [
          "chatops"
        ],
        "summary": "List ChatOps channels",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChatOpsChannelList"
          }
        }
      },
      "post": {
        "tags": [
          "chatops"
        ],
        "summary": "Create a ChatOps channel",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/ChatOpsChannel"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/chatops/messages": {
      "get": {
        "tags": [
          "chatops"
        ],
        "summary": "List ChatOps messages",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChatOpsMessageList"
          }
        }
      }
    },
    "/api/v1/chatops/commands": {
      "post": {
        "tags": [
          "chatops"
        ],
        "summary": "Post a ChatOps command",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChatOpsMessage"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/mobile/sessions": {
      "post": {
        "tags": [
          "mobile"
        ],
        "summary": "Create a mobile session",
        "requestBody": {
          "$ref": "#/components/requestBodies/Object"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nxs_anomaly_session",
        "description": "Browser session cookie issued by /api/v1/auth/login."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key mapped to a role."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same static API key sent as `Authorization: Bearer <key>`."
      },
      "mobileSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Mobile-Session",
        "description": "Mobile session token from /api/v1/mobile/sessions."
      }
    },
...
  }
}